sliptrack/blueprints/uploads/image_utils.py
import cv2
import numpy as np
from PIL import Image

def load_image(path):
    return cv2.imread(path)

def preprocess_for_ocr(path):
    img = cv2.imread(path, cv2.IMREAD_COLOR)
    gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
    gray = cv2.medianBlur(gray, 3)
    thr = cv2.adaptiveThreshold(gray,255,cv2.ADAPTIVE_THRESH_GAUSSIAN_C,
                                cv2.THRESH_BINARY, 31, 15)
    return thr
