requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.0
Flask-Migrate>=4.0.5
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
email-validator>=2.1.0
marshmallow>=3.21.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.9.0.80
python-dotenv>=1.0.1
bcrypt>=4.1.2
flask-limiter>=3.5.0
openpyxl>=3.1.2
XlsxWriter>=3.2.0
pandas>=2.2.0
pytest>=8.0.0
